agentic-ais/atc/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xai-agentic-atc"
version = "0.1.0"
description = "Agentic ATC prototype for X-Plane MCP using AutoGen + GitHub Copilot SDK."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "autogen-agentchat>=0.7.5,<0.8.0",
  "autogen-ext[openai]>=0.7.5,<0.8.0",
  "github-copilot-sdk>=0.1.1,<0.2.0",
  "mcp>=1.13.1,<2.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
]

[project.scripts]
atc-tower = "atc_ai.main:cli_entrypoint"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
